requirements.txt
discord.py~=1.6.0
feedparser~=6.0.2
urllib3~=1.26.2
python-dotenv~=0.17.1